Sahil's Code for KNN and Neural Nets on Datasets
KNN.ipynb contains the algorithms and results of running the knn algorithm on the datasets provided. I specifically ran them on every one except the Digits dataset. 

projModels.ipynb contains all the neural network models for all the datasets except for Digits. 

The notebooks are labeled by dataset, and in the case of the neural networks, each structure is in its own cell, minimized. You can maximize the cell by clicking on the small bar to the left of the cell play button if you are using vscode. 

The outputs and relevant graphs are also present in each notebook.